+++ tests/testthat/test-example_parkinsons.R
@@ -1041,3 +1041,45 @@ test_that("ML-NMR models work with contrast data", {
                                 tolerance = tol)
 })
 
+test_that("Robust to custom options(contrasts) settings", {
+  arm_fit_FE_SAS <- withr::with_options(list(contrasts = c(ordered = "contr.SAS",
+                                                         unordered = "contr.SAS")),
+                  nma(arm_net, 
+                      trt_effects = "fixed",
+                      prior_intercept = normal(scale = 100),
+                      prior_trt = normal(scale = 10),
+                      iter = 10000))
+  
+    expect_equivalent_nma_summary(summary(arm_fit_FE_SAS), summary(arm_fit_FE),
+                                  tolerance = tol)
+    expect_equivalent_nma_summary(relative_effects(arm_fit_FE_SAS), relative_effects(arm_fit_FE),
+                                  tolerance = tol)
+  
+  contr_fit_FE_SAS <- withr::with_options(list(contrasts = c(ordered = "contr.SAS",
+                                                         unordered = "contr.SAS")),
+                  nma(contr_net, 
+                      trt_effects = "fixed",
+                      prior_intercept = normal(scale = 100),
+                      prior_trt = normal(scale = 10),
+                      iter = 10000))
+  
+    expect_equivalent_nma_summary(summary(contr_fit_FE_SAS), summary(contr_fit_FE),
+                                  tolerance = tol)
+    expect_equivalent_nma_summary(relative_effects(contr_fit_FE_SAS), relative_effects(contr_fit_FE),
+                                  tolerance = tol)
+    
+  mix_fit_FE_SAS <- withr::with_options(list(contrasts = c(ordered = "contr.SAS",
+                                                         unordered = "contr.SAS")),
+                  nma(mix_net, 
+                      trt_effects = "fixed",
+                      prior_intercept = normal(scale = 100),
+                      prior_trt = normal(scale = 10),
+                      iter = 10000))
+  
+    expect_equivalent_nma_summary(summary(mix_fit_FE_SAS), summary(mix_fit_FE),
+                                  tolerance = tol)
+    expect_equivalent_nma_summary(relative_effects(mix_fit_FE_SAS), relative_effects(mix_fit_FE),
+                                  tolerance = tol)
+
+})
+